Split out classes and simplify construction of buttons
